Geliştirme
Makaleler.php sayfasında bulunan sorguya ORDER BY eklendi ve
listelemenin id'ye göre yapılması sağlandı. Header php sayfasına toplu
checkbox scripti ve javascript kütüphanesi yerleştirildi. Makaleler.php
sayfasına toplu checkbox eklemesi yapılarak seçili olan makalelerin
silinme işlemi yapıldı ama tamamlanmadı. Sistem üzerinde toplam 2 adet
kullanıcı yetki seviyesi bulunmakta. Sayfalara yetki ataması yapıldı.
Kullanıcıların yetkilendirmesi daha geliştirilebilir.

+++ includes/kullanici_sil.php
@@ -1,5 +1,6 @@
 <?php ob_start(); ?>
 <?php
+if($_SESSION["rol"] == 1){
 	$id = $_GET["c"];
 	$kullanici_sil = mysql_query("DELETE FROM `fozel_blog`.`kullanicilar` WHERE `kullanicilar`.`user_id` = $id",$connection);
 	if(isset($kullanici_sil)){
@@ -9,5 +10,6 @@
 			echo "Bir Hata Oluþtu : ".mysql_error();
 		}
 
+}
 ?>
 <?php ob_end_flush(); ?>
